spliceai/Canonical/requirements.txt
tqdm

torch==2.0.1
numpy==1.23.3
h5py==2.10.0
attrs==20.3.0
pandas==1.3.5
permacache==3.7.0
scipy==1.8.1
scikit-learn==1.0.2
cached-property==1.5.2
matplotlib
matplotlib-venn>=0.11.9
more_itertools
logomaker==0.8
pyensembl==2.2.8
pybbi==0.3.2
parameterized==0.9.0
fire==0.5.0
pulp==2.7.0
